rename Modal to ModalContent
