+++ src/components/UploadCSV/ErrorModalContent/ErrorModalContent.css
