+++ src/components/UploadCSV/UploadModalContent/UploadModalContent.jsx
@@ -2,9 +2,9 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { Button, Text } from '@chakra-ui/react';
 import DropZone from '../DropZone/DropZone';
-import './UploadModal.css';
+import './UploadModalContent.css';
 
-const UploadModal = ({ setCSVFile, onUpload }) => {
+const UploadModalContent = ({ setCSVFile, onUpload }) => {
   return (
     <div className="upload-modal-content">
       <Text className="upload-modal-title">UPLOAD CSV FILE:</Text>
@@ -22,9 +22,9 @@ const UploadModal = ({ setCSVFile, onUpload }) => {
   );
 };
 
-UploadModal.propTypes = {
+UploadModalContent.propTypes = {
   setCSVFile: PropTypes.func.isRequired,
   onUpload: PropTypes.func.isRequired,
 };
 
-export default UploadModal;
+export default UploadModalContent;
